refactoring_finished_task
