+++ src/app/components/qualitiesList.jsx
