+++ src/app/components/searchStatus.jsx
